BAEN-001 getting all the common configuration in library folder
