--- application/.gitignore
@@ -1,15 +1,0 @@
-nbproject
-._*
-.~lock.*
-.buildpath
-.DS_Store
-.idea
-.project
-.settings
-composer.lock
-vendor/bin
-components/
-extras/
-public/static/dojo
-public/static/foundation
-vendor/
