--- application/.gitmodules
@@ -1,3 +1,0 @@
-[submodule "vendor/ZF2"]
-	path = vendor/ZF2
-	url = https://github.com/zendframework/zf2.git
